requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.1.1
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.16.0
cmake==4.2.1
colorama==0.4.6
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
datasets==4.5.0
dill==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.3.1
filelock==3.20.3
frozenlist==1.8.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.0.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langcodes==3.5.1
lightning-utilities==0.15.3
lit==18.1.8
lxml==6.0.2
Markdown==3.10.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.7.1
multiprocess==0.70.15
murmurhash==1.0.15
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pathtools==0.1.2
pillow==12.1.1
portalocker==3.2.0
preshed==3.0.12
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
sacrebleu==2.6.0
sentry-sdk==2.53.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
thinc==8.2.5
tokenizers==0.13.3
toolz==1.1.0
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.0.3
torchtext==0.15.2
torchvision==0.15.2
tqdm==4.67.3
triton==2.0.0
typer==0.9.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.15.9
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.5
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
